requirements.txt
# Generated from requirements.in
# To update: pip-compile requirements.in

annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distro==1.9.0
lxml==5.3.0
fastapi==0.115.5
greenlet==3.1.1
grpcio==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
jiter==0.7.1
mypy==1.13.0
mypy-extensions==1.0.0
openai==1.54.5
packaging==24.2
pathspec==0.12.1
pdfplumber==0.11.4
pillow==11.0.0
platformdirs==4.3.6
portalocker==2.10.1
protobuf==5.28.3
pycparser==2.22
pydantic==2.10.2
pydantic-core==2.27.1
pydantic-settings==2.6.1
pymysql==1.1.1
pypdf==5.1.0
pypdfium2==4.30.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
pyyaml==6.0.2
qdrant-client==1.12.1
requests==2.32.3
ruff==0.8.0
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.41.3
structlog==24.4.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.0
trafilatura==1.12.2
types-pyyaml==6.0.12.20240917
types-requests==2.32.0.20241016
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==0.24.0
websockets==14.1

